LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.fontStretch.settings.worker-expected.txt


FAIL Testing value setting of fontStretch in Canvas assert_equals: ctx.fontStretch === "ultra-expanded" (got ulTra-condensed[string], expected ultra-expanded[string]) expected "ultra-expanded" but got "ulTra-condensed"
